feat: create downloaders test_utils mod
 * add test utils feature
