+++ crates/net/downloaders/src/lib.rs
@@ -17,5 +17,5 @@ pub mod headers;
 /// Common downloader metrics.
 pub mod metrics;
 
-#[cfg(test)]
-mod test_utils;
+#[cfg(any(test, feature = "test-utils"))]
+pub mod test_utils;
